requirements.txt
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
pydantic>=2.5.2
python-dotenv>=1.0.0
openai>=1.3.7
requests>=2.31.0
PyMuPDF>=1.23.8
sentence-transformers>=2.2.2
faiss-cpu>=1.9.0
langchain>=0.2.1
numpy>=1.26.0

